basic restructuring of ElectronicTransition, ExcitedStates and State2States
